Refactor to Send a webhook Moved to WebhookBundle
