+++ app/bundles/CampaignBundle/Translations/en_US/messages.ini
@@ -15,14 +15,6 @@ mautic.campaign.event.action.header="Action"
 mautic.campaign.event.actions.header="Actions"
 mautic.campaign.event.addremovelead="Change campaigns"
 mautic.campaign.event.addremovelead_descr="Add contact to specific campaigns and/or remove from specific campaigns when the event is triggered."
-mautic.campaign.event.remoteurl="Send a webhook"
-mautic.campaign.event.remoteurl_desc="Send a webhook (only for experienced users)"
-mautic.campaign.campaign.remoteurl="Send a webhook"
-mautic.campaign.event.remoteurl.url="Url"
-mautic.campaign.event.remoteurl.method="Method"
-mautic.campaign.event.remoteurl.data="Data"
-mautic.campaign.event.remoteurl.headers="Headers"
-mautic.campaign.event.remoteurl.timeout="Timeout"
 mautic.campaign.event.condition.header="Condition"
 mautic.campaign.event.condition.descr="A condition is based on known profile field values or submitted form data."
 mautic.campaign.event.conditions.header="Conditions"
