+++ app/bundles/WebhookBundle/Translations/en_US/messages.ini
@@ -34,3 +34,11 @@ mautic.webhook.runtime="Request Runtime"
 mautic.webhook.killed="Webhook was automatically unpublished, because the response has errored out %limit% times in a row."
 mautic.notification.permissions.header="Notification Permissions"
 mautic.notification.permissions.notifications="Notifications - User has access to"
+mautic.webhook.event.sendwebhook="Send a webhook"
+mautic.webhook.event.sendwebhook_desc="Send a webhook (only for experienced users)"
+mautic.campaign.campaign.sendwebhook="Send a webhook"
+mautic.webhook.event.sendwebhook.url="Url"
+mautic.webhook.event.sendwebhook.method="Method"
+mautic.webhook.event.sendwebhook.data="Data"
+mautic.webhook.event.sendwebhook.headers="Headers"
+mautic.webhook.event.sendwebhook.timeout="Timeout"
